+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -1288,10 +1288,6 @@ class GeckoEngineSession(
         override fun onShowDynamicToolbar(geckoSession: GeckoSession) {
             notifyObservers { onShowDynamicToolbar() }
         }
-
-        override fun onGetNimbusFeature(session: GeckoSession, featureId: String): JSONObject? {
-            return GeckoNimbus.getFeature(featureId)?.toJSONObject()
-        }
     }
 
     private fun createContentBlockingDelegate() = object : ContentBlocking.Delegate {