+++ mobile/android/android-components/samples/browser/src/gecko/java/org/mozilla/samples/browser/Components.kt
@@ -8,6 +8,7 @@ import mozilla.components.browser.engine.gecko.GeckoEngine
 import mozilla.components.browser.engine.gecko.fetch.GeckoViewFetchClient
 import mozilla.components.browser.engine.gecko.glean.GeckoAdapter
 import mozilla.components.concept.engine.Engine
+import mozilla.components.experiment.NimbusExperimentDelegate
 import mozilla.components.feature.webcompat.WebCompatFeature
 import mozilla.components.feature.webcompat.reporter.WebCompatReporterFeature
 import mozilla.components.lib.crash.handler.CrashHandlerService
@@ -23,6 +24,7 @@ class Components(private val applicationContext: Context) : DefaultComponents(ap
         // Allow for exfiltrating Gecko metrics through the Glean SDK.
         val builder = GeckoRuntimeSettings.Builder().aboutConfigEnabled(true)
         builder.telemetryDelegate(GeckoAdapter())
+        builder.experimentDelegate(NimbusExperimentDelegate())
         builder.crashHandler(CrashHandlerService::class.java)
         GeckoRuntime.create(applicationContext, builder.build())
     }